Split hostspool_mgr.go in multiple files to be more readable
