+++ prov/hostspool/hostspool_mgr.go
@@ -28,11 +28,9 @@ import (
 	"github.com/pkg/errors"
 	"golang.org/x/crypto/ssh"
 
-	"github.com/ystia/yorc/config"
 	"github.com/ystia/yorc/helper/consulutil"
 	"github.com/ystia/yorc/helper/labelsutil"
 	"github.com/ystia/yorc/helper/sshutil"
-	"github.com/ystia/yorc/log"
 )
 
 const (
@@ -54,7 +52,7 @@ type Manager interface {
 	UpdateLabels(hostname string, diff map[string]string, operation func(a int64, b int64) int64, update func(orig map[string]string, diff map[string]string, operation func(a int64, b int64) int64) (map[string]string, error)) error
 	AddLabels(hostname string, labels map[string]string) error
 	RemoveLabels(hostname string, labels []string) error
-	UpdateHost(hostname string, connection Connection) error
+	UpdateConnection(hostname string, connection Connection) error
 	List(filters ...labelsutil.Filter) ([]string, []labelsutil.Warning, uint64, error)
 	GetHost(hostname string) (Host, error)
 	Allocate(allocation *Allocation, filters ...labelsutil.Filter) (string, []labelsutil.Warning, error)
@@ -98,7 +96,6 @@ func (cm *consulManager) Add(hostname string, conn Connection, labels map[string
 	return cm.addWait(hostname, conn, labels, maxWaitTimeSeconds*time.Second)
 }
 func (cm *consulManager) addWait(hostname string, conn Connection, labels map[string]string, maxWaitTime time.Duration) error {
-
 	ops, err := getAddOperations(hostname, conn, labels, HostStatusFree, "", nil)
 	if err != nil {
 		return err
@@ -231,226 +228,6 @@ func getAddOperations(
 	return addOps, nil
 }
 
-func getAddAllocationsOperation(hostname string, allocations []Allocation) (api.KVTxnOps, error) {
-	allocsOps := api.KVTxnOps{}
-	hostKVPrefix := path.Join(consulutil.HostsPoolPrefix, hostname)
-	if allocations != nil {
-		for _, alloc := range allocations {
-			allocKVPrefix := path.Join(hostKVPrefix, "allocations", alloc.ID)
-			allocOps := api.KVTxnOps{
-				&api.KVTxnOp{
-					Verb:  api.KVSet,
-					Key:   path.Join(allocKVPrefix),
-					Value: []byte(alloc.ID),
-				},
-				&api.KVTxnOp{
-					Verb:  api.KVSet,
-					Key:   path.Join(allocKVPrefix, "node_name"),
-					Value: []byte(alloc.NodeName),
-				},
-				&api.KVTxnOp{
-					Verb:  api.KVSet,
-					Key:   path.Join(allocKVPrefix, "instance"),
-					Value: []byte(alloc.Instance),
-				},
-				&api.KVTxnOp{
-					Verb:  api.KVSet,
-					Key:   path.Join(allocKVPrefix, "deployment_id"),
-					Value: []byte(alloc.DeploymentID),
-				},
-				&api.KVTxnOp{
-					Verb:  api.KVSet,
-					Key:   path.Join(allocKVPrefix, "shareable"),
-					Value: []byte(strconv.FormatBool(alloc.Shareable)),
-				},
-			}
-
-			for k, v := range alloc.Resources {
-				k = url.PathEscape(k)
-				if k == "" {
-					return nil, errors.WithStack(badRequestError{"empty labels are not allowed"})
-				}
-				allocOps = append(allocOps, &api.KVTxnOp{
-					Verb:  api.KVSet,
-					Key:   path.Join(allocKVPrefix, "resources", k),
-					Value: []byte(v),
-				})
-			}
-
-			allocsOps = append(allocsOps, allocOps...)
-		}
-	}
-	return allocsOps, nil
-}
-
-func (cm *consulManager) UpdateHost(hostname string, conn Connection) error {
-	return cm.updateHostWait(hostname, conn, maxWaitTimeSeconds*time.Second)
-}
-func (cm *consulManager) updateHostWait(hostname string, conn Connection, maxWaitTime time.Duration) error {
-	if hostname == "" {
-		return errors.WithStack(badRequestError{`"hostname" missing`})
-	}
-
-	// check if host exists
-	status, err := cm.GetHostStatus(hostname)
-	if err != nil {
-		return err
-	}
-
-	ops := make(api.KVTxnOps, 0)
-	hostKVPrefix := path.Join(consulutil.HostsPoolPrefix, hostname)
-	if conn.User != "" {
-		ops = append(ops, &api.KVTxnOp{
-			Verb:  api.KVSet,
-			Key:   path.Join(hostKVPrefix, "connection", "user"),
-			Value: []byte(conn.User),
-		})
-	}
-	if conn.Port != 0 {
-		ops = append(ops, &api.KVTxnOp{
-			Verb:  api.KVSet,
-			Key:   path.Join(hostKVPrefix, "connection", "port"),
-			Value: []byte(strconv.FormatUint(conn.Port, 10)),
-		})
-	}
-	if conn.Host != "" {
-		ops = append(ops, &api.KVTxnOp{
-			Verb:  api.KVSet,
-			Key:   path.Join(hostKVPrefix, "connection", "host"),
-			Value: []byte(conn.Host),
-		})
-	}
-	if conn.PrivateKey != "" {
-		if conn.PrivateKey == "-" {
-			ok, err := cm.DoesHostHasConnectionPassword(hostname)
-			if err != nil {
-				return err
-			}
-			if !ok && conn.Password == "" || ok && conn.Password == "-" {
-				return errors.WithStack(badRequestError{`at any time at least one of "password" or "private_key" is required`})
-			}
-			conn.PrivateKey = ""
-		}
-		ops = append(ops, &api.KVTxnOp{
-			Verb:  api.KVSet,
-			Key:   path.Join(hostKVPrefix, "connection", "private_key"),
-			Value: []byte(conn.PrivateKey),
-		})
-	}
-	if conn.Password != "" {
-		if conn.Password == "-" {
-			ok, err := cm.DoesHostHasConnectionPrivateKey(hostname)
-			if err != nil {
-				return err
-			}
-			if !ok && conn.PrivateKey == "" || ok && conn.PrivateKey == "-" {
-				return errors.WithStack(badRequestError{`at any time at least one of "password" or "private_key" is required`})
-			}
-			conn.Password = ""
-		}
-		ops = append(ops, &api.KVTxnOp{
-			Verb:  api.KVSet,
-			Key:   path.Join(hostKVPrefix, "connection", "password"),
-			Value: []byte(conn.Password),
-		})
-	}
-
-	_, cleanupFn, err := cm.lockKey(hostname, "update", maxWaitTime)
-	if err != nil {
-		return err
-	}
-	defer cleanupFn()
-
-	ok, response, _, err := cm.cc.KV().Txn(ops, nil)
-	if err != nil {
-		return errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-	}
-	if !ok {
-		// Check the response
-		errs := make([]string, 0)
-		for _, e := range response.Errors {
-			errs = append(errs, e.What)
-		}
-		return errors.Errorf("Failed to update host %q connection: %s", hostname, strings.Join(errs, ", "))
-	}
-
-	err = cm.checkConnection(hostname)
-	if err != nil {
-		if status != HostStatusError {
-			cm.backupHostStatus(hostname)
-			cm.setHostStatusWithMessage(hostname, HostStatusError, "failed to connect to host")
-		}
-		return err
-	}
-	if status == HostStatusError {
-		cm.restoreHostStatus(hostname)
-	}
-	return nil
-}
-
-func (cm *consulManager) UpdateLabels(hostname string, diff map[string]string, operation func(a int64, b int64) int64, update func(orig map[string]string, diff map[string]string, operation func(a int64, b int64) int64) (map[string]string, error)) error {
-	return cm.updateLabelsWait(hostname, diff, operation, update, maxWaitTimeSeconds*time.Second)
-}
-
-func (cm *consulManager) updateLabelsWait(hostname string, diff map[string]string, operation func(a int64, b int64) int64, update func(orig map[string]string, diff map[string]string, operation func(a int64, b int64) int64) (map[string]string, error), maxWaitTime time.Duration) error {
-	if hostname == "" {
-		return errors.WithStack(badRequestError{`"hostname" missing`})
-	}
-
-	lockCh, cleanupFn, err := cm.lockKey(hostname, "updateLabels", maxWaitTime)
-	if err != nil {
-		return err
-	}
-	defer cleanupFn()
-
-	select {
-	case <-lockCh:
-		return errors.Errorf("admin lock lost on hosts pool for updating labels with host %q", hostname)
-	default:
-	}
-
-	labels, err := cm.GetHostLabels(hostname)
-
-	upLabels, err := update(labels, diff, operation)
-	if err != nil {
-		return err
-	}
-
-	if upLabels == nil || len(upLabels) == 0 {
-		return nil
-	}
-
-	log.Debugf("Updating labels:%+v", upLabels)
-	return cm.addLabels(hostname, upLabels)
-}
-
-func (cm *consulManager) getUpdateLabelsOperations(hostname string, diff map[string]string, new map[string]string, operation func(a int64, b int64) int64, update func(orig map[string]string, diff map[string]string, operation func(a int64, b int64) int64) (map[string]string, error)) (api.KVTxnOps, error) {
-	upLabels, err := update(new, diff, operation)
-	if err != nil {
-		return nil, err
-	}
-
-	if upLabels == nil || len(upLabels) == 0 {
-		return nil, nil
-	}
-
-	hostKVPrefix := path.Join(consulutil.HostsPoolPrefix, hostname)
-	ops := make(api.KVTxnOps, 0)
-
-	for k, v := range upLabels {
-		k = url.PathEscape(k)
-		if k == "" {
-			return nil, errors.WithStack(badRequestError{"empty labels are not allowed"})
-		}
-		ops = append(ops, &api.KVTxnOp{
-			Verb:  api.KVSet,
-			Key:   path.Join(hostKVPrefix, "labels", k),
-			Value: []byte(v),
-		})
-	}
-	return ops, nil
-}
-
 func (cm *consulManager) Remove(hostname string) error {
 	return cm.removeWait(hostname, maxWaitTimeSeconds*time.Second)
 }
@@ -517,119 +294,6 @@ func (cm *consulManager) getRemoveOperations(hostname string, checkStatus bool) 
 	}
 
 	return rmOps, nil
-}
-
-func (cm *consulManager) addLabels(hostname string, labels map[string]string) error {
-	hostKVPrefix := path.Join(consulutil.HostsPoolPrefix, hostname)
-	ops := make(api.KVTxnOps, 0)
-
-	for k, v := range labels {
-		k = url.PathEscape(k)
-		if k == "" {
-			return errors.WithStack(badRequestError{"empty labels are not allowed"})
-		}
-		ops = append(ops, &api.KVTxnOp{
-			Verb:  api.KVSet,
-			Key:   path.Join(hostKVPrefix, "labels", k),
-			Value: []byte(v),
-		})
-	}
-	ok, response, _, err := cm.cc.KV().Txn(ops, nil)
-	if err != nil {
-		return errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-	}
-	if !ok {
-		// Check the response
-		errs := make([]string, 0)
-		for _, e := range response.Errors {
-			errs = append(errs, e.What)
-		}
-		return errors.Errorf("Failed to add labels to host %q: %s", hostname, strings.Join(errs, ", "))
-	}
-
-	return nil
-}
-
-func (cm *consulManager) AddLabels(hostname string, labels map[string]string) error {
-	return cm.addLabelsWait(hostname, labels, maxWaitTimeSeconds*time.Second)
-}
-func (cm *consulManager) addLabelsWait(hostname string, labels map[string]string, maxWaitTime time.Duration) error {
-	if hostname == "" {
-		return errors.WithStack(badRequestError{`"hostname" missing`})
-	}
-	if labels == nil || len(labels) == 0 {
-		return nil
-	}
-
-	_, cleanupFn, err := cm.lockKey(hostname, "labels addition", maxWaitTime)
-	if err != nil {
-		return err
-	}
-	defer cleanupFn()
-
-	// Checks host existence
-	// We don't care about host status for updating labels
-	_, err = cm.GetHostStatus(hostname)
-	if err != nil {
-		return err
-	}
-
-	return cm.addLabels(hostname, labels)
-}
-
-func (cm *consulManager) RemoveLabels(hostname string, labels []string) error {
-	return cm.removeLabelsWait(hostname, labels, maxWaitTimeSeconds*time.Second)
-}
-func (cm *consulManager) removeLabelsWait(hostname string, labels []string, maxWaitTime time.Duration) error {
-	if hostname == "" {
-		return errors.WithStack(badRequestError{`"hostname" missing`})
-	}
-	if labels == nil || len(labels) == 0 {
-		return nil
-	}
-
-	hostKVPrefix := path.Join(consulutil.HostsPoolPrefix, hostname)
-	ops := make(api.KVTxnOps, 0)
-
-	for _, v := range labels {
-		v = url.PathEscape(v)
-		if v == "" {
-			return errors.WithStack(badRequestError{"empty labels are not allowed"})
-		}
-		ops = append(ops, &api.KVTxnOp{
-			Verb: api.KVDelete,
-			Key:  path.Join(hostKVPrefix, "labels", v),
-		})
-	}
-
-	_, cleanupFn, err := cm.lockKey(hostname, "labels remove", maxWaitTime)
-	if err != nil {
-		return err
-	}
-	defer cleanupFn()
-
-	// Checks host existence
-	_, err = cm.GetHostStatus(hostname)
-	if err != nil {
-		return err
-	}
-
-	// We don't care about host status for updating labels
-
-	ok, response, _, err := cm.cc.KV().Txn(ops, nil)
-	if err != nil {
-		return errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-	}
-	if !ok {
-		// Check the response
-		errs := make([]string, 0)
-		for _, e := range response.Errors {
-			errs = append(errs, e.What)
-		}
-		return errors.Errorf("Failed to delete labels on host %q: %s", hostname, strings.Join(errs, ", "))
-	}
-
-	return nil
 }
 
 func (cm *consulManager) lockKey(hostname, opType string, lockWaitTime time.Duration) (lockCh <-chan struct{}, cleanupFn func(), err error) {
@@ -820,175 +484,6 @@ func (cm *consulManager) getStatus(hostname string, backup bool) (HostStatus, er
 	return status, nil
 }
 
-func (cm *consulManager) addAllocation(hostname string, allocation *Allocation) error {
-	var allocOps api.KVTxnOps
-	var err error
-	if allocOps, err = getAddAllocationsOperation(hostname, []Allocation{*allocation}); err != nil {
-		return errors.Wrapf(err, "failed to add allocation to host:%q", hostname)
-	}
-
-	ok, response, _, err := cm.cc.KV().Txn(allocOps, nil)
-	if err != nil {
-		return errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-	}
-	if !ok {
-		// Check the response
-		errs := make([]string, 0)
-		for _, e := range response.Errors {
-			errs = append(errs, e.What)
-		}
-		return errors.Errorf("Failed to add allocation on host %q: %s", hostname, strings.Join(errs, ", "))
-	}
-	return nil
-}
-
-func (cm *consulManager) removeAllocation(hostname string, allocation *Allocation) error {
-	_, err := cm.cc.KV().DeleteTree(path.Join(consulutil.HostsPoolPrefix, hostname, "allocations", allocation.ID), nil)
-	return errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-}
-
-func exist(allocations []Allocation, ID string) bool {
-	for _, alloc := range allocations {
-		if alloc.ID == ID {
-			return true
-		}
-	}
-	return false
-}
-
-func (cm *consulManager) GetAllocations(hostname string) ([]Allocation, error) {
-	allocations := make([]Allocation, 0)
-	if hostname == "" {
-		return nil, errors.WithStack(badRequestError{`"hostname" missing`})
-	}
-	keys, _, err := cm.cc.KV().Keys(path.Join(consulutil.HostsPoolPrefix, hostname, "allocations")+"/", "/", nil)
-	if err != nil {
-		return nil, errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-	}
-
-	for _, key := range keys {
-		id := path.Base(key)
-		if exist(allocations, id) {
-			continue
-		}
-		alloc := Allocation{}
-		alloc.ID = id
-		kvp, _, err := cm.cc.KV().Get(path.Join(key, "node_name"), nil)
-		if err != nil {
-			return nil, errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-		}
-		if kvp != nil && len(kvp.Value) > 0 {
-			alloc.NodeName = string(kvp.Value)
-		}
-
-		kvp, _, err = cm.cc.KV().Get(path.Join(key, "instance"), nil)
-		if err != nil {
-			return nil, errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-		}
-		if kvp != nil && len(kvp.Value) > 0 {
-			alloc.Instance = string(kvp.Value)
-		}
-
-		kvp, _, err = cm.cc.KV().Get(path.Join(key, "deployment_id"), nil)
-		if err != nil {
-			return nil, errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-		}
-		if kvp != nil && len(kvp.Value) > 0 {
-			alloc.DeploymentID = string(kvp.Value)
-		}
-
-		kvp, _, err = cm.cc.KV().Get(path.Join(key, "shareable"), nil)
-		if err != nil {
-			return nil, errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-		}
-		if kvp != nil && len(kvp.Value) > 0 {
-			alloc.Shareable, err = strconv.ParseBool(string(kvp.Value))
-			if err != nil {
-				return nil, errors.Wrapf(err, "failed to parse boolean from value:%q", string(kvp.Value))
-			}
-		}
-
-		kvps, _, err := cm.cc.KV().List(path.Join(key, "resources"), nil)
-		if err != nil {
-			return nil, errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-		}
-		resources := make(map[string]string, len(kvps))
-		for _, kvp := range kvps {
-			resources[path.Base(kvp.Key)] = string(kvp.Value)
-		}
-		alloc.Resources = resources
-
-		allocations = append(allocations, alloc)
-	}
-	return allocations, nil
-}
-
-func (cm *consulManager) DoesHostHasConnectionPrivateKey(hostname string) (bool, error) {
-	c, err := cm.GetHostConnection(hostname)
-	if err != nil {
-		return false, err
-	}
-	return c.PrivateKey != "", nil
-}
-
-func (cm *consulManager) DoesHostHasConnectionPassword(hostname string) (bool, error) {
-	c, err := cm.GetHostConnection(hostname)
-	if err != nil {
-		return false, err
-	}
-	return c.Password != "", nil
-}
-
-func (cm *consulManager) GetHostConnection(hostname string) (Connection, error) {
-	conn := Connection{}
-	if hostname == "" {
-		return conn, errors.WithStack(badRequestError{`"hostname" missing`})
-	}
-	kv := cm.cc.KV()
-	connKVPrefix := path.Join(consulutil.HostsPoolPrefix, hostname, "connection")
-
-	kvp, _, err := kv.Get(path.Join(connKVPrefix, "host"), nil)
-	if err != nil {
-		return conn, errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-	}
-	if kvp != nil {
-		conn.Host = string(kvp.Value)
-	}
-	kvp, _, err = kv.Get(path.Join(connKVPrefix, "user"), nil)
-	if err != nil {
-		return conn, errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-	}
-	if kvp != nil {
-		conn.User = string(kvp.Value)
-	}
-	kvp, _, err = kv.Get(path.Join(connKVPrefix, "password"), nil)
-	if err != nil {
-		return conn, errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-	}
-	if kvp != nil {
-		conn.Password = string(kvp.Value)
-	}
-	kvp, _, err = kv.Get(path.Join(connKVPrefix, "private_key"), nil)
-	if err != nil {
-		return conn, errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-	}
-	if kvp != nil {
-		conn.PrivateKey = string(kvp.Value)
-	}
-	kvp, _, err = kv.Get(path.Join(connKVPrefix, "port"), nil)
-	if err != nil {
-		return conn, errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-	}
-	if kvp != nil {
-		conn.Port, err = strconv.ParseUint(string(kvp.Value), 10, 64)
-		if err != nil {
-			return conn, errors.Wrapf(err, "failed to retrieve connection port for host %q", hostname)
-		}
-	}
-
-	return conn, nil
-}
-
 func (cm *consulManager) GetHostMessage(hostname string) (string, error) {
 	return cm.getMessage(hostname, false)
 }
@@ -1031,26 +526,6 @@ func (cm *consulManager) setHostMessage(hostname, message string) error {
 	return consulutil.StoreConsulKeyAsString(path.Join(consulutil.HostsPoolPrefix, hostname, "message"), message)
 }
 
-func (cm *consulManager) GetHostLabels(hostname string) (map[string]string, error) {
-	if hostname == "" {
-		return nil, errors.WithStack(badRequestError{`"hostname" missing`})
-	}
-	// check if host exists
-	_, err := cm.GetHostStatus(hostname)
-	if err != nil {
-		return nil, err
-	}
-	kvps, _, err := cm.cc.KV().List(path.Join(consulutil.HostsPoolPrefix, hostname, "labels"), nil)
-	if err != nil {
-		return nil, errors.Wrap(err, consulutil.ConsulGenericErrMsg)
-	}
-	labels := make(map[string]string, len(kvps))
-	for _, kvp := range kvps {
-		labels[path.Base(kvp.Key)] = string(kvp.Value)
-	}
-	return labels, nil
-}
-
 func (cm *consulManager) GetHost(hostname string) (Host, error) {
 	host := Host{Name: hostname}
 	if hostname == "" {
@@ -1077,142 +552,6 @@ func (cm *consulManager) GetHost(hostname string) (Host, error) {
 
 	host.Labels, err = cm.GetHostLabels(hostname)
 	return host, err
-}
-
-func (cm *consulManager) Allocate(allocation *Allocation, filters ...labelsutil.Filter) (string, []labelsutil.Warning, error) {
-	return cm.allocateWait(maxWaitTimeSeconds*time.Second, allocation, filters...)
-}
-func (cm *consulManager) allocateWait(maxWaitTime time.Duration, allocation *Allocation, filters ...labelsutil.Filter) (string, []labelsutil.Warning, error) {
-	// Build allocationID
-	if err := allocation.buildID(); err != nil {
-		return "", nil, err
-	}
-
-	lockCh, cleanupFn, err := cm.lockKey("", "allocation", maxWaitTime)
-	if err != nil {
-		return "", nil, err
-	}
-	defer cleanupFn()
-
-	hosts, warnings, _, err := cm.List(filters...)
-	if err != nil {
-		return "", warnings, err
-	}
-	// Filters only free or allocated hosts in case of shareable allocation
-	var lastErr error
-	freeHosts := hosts[:0]
-	for _, h := range hosts {
-		select {
-		case <-lockCh:
-			return "", warnings, errors.New("admin lock lost on hosts pool during host allocation")
-		default:
-		}
-		err := cm.checkConnection(h)
-		if err != nil {
-			lastErr = err
-			continue
-		}
-		hs, err := cm.GetHostStatus(h)
-		if err != nil {
-			lastErr = err
-		} else {
-			if hs == HostStatusFree {
-				freeHosts = append(freeHosts, h)
-			} else if hs == HostStatusAllocated && allocation.Shareable {
-				allocations, err := cm.GetAllocations(h)
-				if err != nil {
-					lastErr = err
-					continue
-				}
-				// Check the host allocation is not unshareable
-				if len(allocations) == 1 && !allocations[0].Shareable {
-					continue
-				}
-				freeHosts = append(freeHosts, h)
-			}
-		}
-	}
-
-	if len(freeHosts) == 0 {
-		if lastErr != nil {
-			return "", warnings, lastErr
-		}
-		return "", warnings, errors.WithStack(noMatchingHostFoundError{})
-	}
-	// Get the first host that match
-	hostname := freeHosts[0]
-	select {
-	case <-lockCh:
-		return "", warnings, errors.New("admin lock lost on hosts pool during host allocation")
-	default:
-	}
-
-	if err := cm.addAllocation(hostname, allocation); err != nil {
-		return "", warnings, errors.Wrapf(err, "failed to add allocation for hostname:%q", hostname)
-	}
-
-	return hostname, warnings, cm.setHostStatus(hostname, HostStatusAllocated)
-}
-func (cm *consulManager) Release(hostname string, allocation *Allocation) error {
-	return cm.releaseWait(hostname, allocation, maxWaitTimeSeconds*time.Second)
-}
-
-func (cm *consulManager) releaseWait(hostname string, allocation *Allocation, maxWaitTime time.Duration) error {
-	// Build allocationID
-	if err := allocation.buildID(); err != nil {
-		return err
-	}
-	_, cleanupFn, err := cm.lockKey(hostname, "release", maxWaitTime)
-	if err != nil {
-		return err
-	}
-	defer cleanupFn()
-
-	if err := cm.removeAllocation(hostname, allocation); err != nil {
-		return errors.Wrapf(err, "failed to remove allocation with ID:%q and hostname:%q", allocation.ID, hostname)
-	}
-
-	host, err := cm.GetHost(hostname)
-	if err != nil {
-		return err
-	}
-	// Set the host status to free only for host with no allocations
-	if len(host.Allocations) == 0 {
-		if err = cm.setHostStatus(hostname, HostStatusFree); err != nil {
-			return err
-		}
-	}
-	err = cm.checkConnection(hostname)
-	if err != nil {
-		cm.backupHostStatus(hostname)
-		cm.setHostStatusWithMessage(hostname, HostStatusError, "failed to connect to host")
-	}
-	return nil
-}
-
-func resolveTemplatesInConnection(conn *Connection) {
-	conn.User = config.DefaultConfigTemplateResolver.ResolveValueWithTemplates("Connection.User", conn.User).(string)
-	conn.Password = config.DefaultConfigTemplateResolver.ResolveValueWithTemplates("Connection.Password", conn.Password).(string)
-	conn.PrivateKey = config.DefaultConfigTemplateResolver.ResolveValueWithTemplates("Connection.PrivateKey", conn.PrivateKey).(string)
-	conn.Host = config.DefaultConfigTemplateResolver.ResolveValueWithTemplates("Connection.Host", conn.Host).(string)
-}
-
-// Check if we can log into an host given a connection
-func (cm *consulManager) checkConnection(hostname string) error {
-
-	conn, err := cm.GetHostConnection(hostname)
-	if err != nil {
-		return errors.Wrapf(err, "failed to connect to host %q", hostname)
-	}
-	resolveTemplatesInConnection(&conn)
-	conf, err := getSSHConfig(conn)
-	if err != nil {
-		return errors.Wrapf(err, "failed to connect to host %q", hostname)
-	}
-
-	client := cm.getSSHClient(conf, conn)
-	_, err = client.RunCommand(`echo "Connected!"`)
-	return errors.Wrapf(err, "failed to connect to host %q", hostname)
 }
 
 func getSSHConfig(conn Connection) (*ssh.ClientConfig, error) {
@@ -1456,28 +795,4 @@ func (cm *consulManager) applyWait(
 	}
 
 	return err
-}
-
-// Go routine checking a Host connection and updating the Host status
-func (cm *consulManager) updateConnectionStatus(name string, waitGroup *sync.WaitGroup) {
-	defer waitGroup.Done()
-	status, err := cm.GetHostStatus(name)
-	if err != nil {
-		// No such host anymore
-		return
-	}
-
-	err = cm.checkConnection(name)
-	if err != nil {
-		if status != HostStatusError {
-			cm.backupHostStatus(name)
-			cm.setHostStatusWithMessage(name, HostStatusError, "failed to connect to host")
-		}
-		return
-	}
-	// Connection is up now. If it was previously down, restoring the status as
-	// it was before the failure (free, allocated)
-	if status == HostStatusError {
-		cm.restoreHostStatus(name)
-	}
 }
